+++ spmv/src/spmv_cuda.cu
@@ -11,10 +11,10 @@
 
 #include <sys/time.h>
 
-SmdvInfo init_mul (size_t num_rows, size_t num_cols, size_t val_num);
+SmdvInfo init_mul (int num_rows, int num_cols, int val_num);
 float exec_mul (
-    float *dst, size_t *row, size_t *col, float *val, float *arr,
-    size_t num_rows, size_t num_cols, size_t val_num, SmdvInfo& smdv_info
+    float *dst, int *row, int *col, float *val, float *arr,
+    int num_rows, int num_cols, int val_num, SmdvInfo& smdv_info
 );
 void deinit_mul (SmdvInfo &smdv_info);
 
@@ -22,7 +22,7 @@ void deinit_mul (SmdvInfo &smdv_info);
 
 __global__
 void mul_kernel (
-    float *dst, size_t *row, size_t *col, float *val, float *arr, size_t val_num
+    float *dst, int *row, int *col, float *val, float *arr, int val_num
 )
 {
     int idx = blockIdx.x * blockDim.x + threadIdx.x;
@@ -39,8 +39,8 @@ void mul_kernel (
 
 
 float mul_cuda (
-    float *dst, size_t *row, size_t *col, float *val, float *arr,
-    size_t num_rows, size_t num_cols, size_t val_num
+    float *dst, int *row, int *col, float *val, float *arr,
+    int num_rows, int num_cols, int val_num
 )
 {
     auto smdv_info = init_mul(num_rows, num_cols, val_num);
@@ -52,15 +52,15 @@ float mul_cuda (
 
 
 float exec_mul (
-    float *dst, size_t *row, size_t *col, float *val, float *arr,
-    size_t num_rows, size_t num_cols, size_t val_num, SmdvInfo& smdv_info
+    float *dst, int *row, int *col, float *val, float *arr,
+    int num_rows, int num_cols, int val_num, SmdvInfo& smdv_info
 ) 
 {
     // Memcpy
     SAFE_CALL( cudaMemcpy(smdv_info.d_val, val, sizeof(float) * val_num, cudaMemcpyHostToDevice) );
     SAFE_CALL( cudaMemcpy(smdv_info.d_arr, arr, sizeof(float) * num_cols, cudaMemcpyHostToDevice) );
-    SAFE_CALL( cudaMemcpy(smdv_info.d_row, row, sizeof(size_t) * val_num, cudaMemcpyHostToDevice) );
-    SAFE_CALL( cudaMemcpy(smdv_info.d_col, col, sizeof(size_t) * val_num, cudaMemcpyHostToDevice) );
+    SAFE_CALL( cudaMemcpy(smdv_info.d_row, row, sizeof(int) * val_num, cudaMemcpyHostToDevice) );
+    SAFE_CALL( cudaMemcpy(smdv_info.d_col, col, sizeof(int) * val_num, cudaMemcpyHostToDevice) );
 
     dim3 dim_grid = dim3(smdv_info.dim.grid.x, smdv_info.dim.grid.y, smdv_info.dim.grid.z);
     dim3 dim_block = dim3(smdv_info.dim.block.x, smdv_info.dim.block.y, smdv_info.dim.block.z);
@@ -94,7 +94,7 @@ float exec_mul (
 }
 
 
-SmdvInfo init_mul (size_t num_rows, size_t num_cols, size_t val_num)
+SmdvInfo init_mul (int num_rows, int num_cols, int val_num)
 {
     SmdvInfo smdv_info;
 
@@ -103,10 +103,10 @@ SmdvInfo init_mul (size_t num_rows, size_t num_cols, size_t val_num)
     SAFE_CALL( cudaMemset(smdv_info.d_dst, 0, sizeof(float) * num_rows) );
     SAFE_CALL( cudaMalloc(&smdv_info.d_val, sizeof(float) * val_num) );
     SAFE_CALL( cudaMalloc(&smdv_info.d_arr, sizeof(float) * num_cols) );
-    SAFE_CALL( cudaMalloc(&smdv_info.d_row, sizeof(size_t) * val_num) );
-    SAFE_CALL( cudaMalloc(&smdv_info.d_col, sizeof(size_t) * val_num) );
+    SAFE_CALL( cudaMalloc(&smdv_info.d_row, sizeof(int) * val_num) );
+    SAFE_CALL( cudaMalloc(&smdv_info.d_col, sizeof(int) * val_num) );
 
-    find_best_grid_linear(smdv_info.dim, val_num);
+    find_best_grid_linear(smdv_info.dim, size_t(val_num));
 
     return smdv_info;
 }
